dark/light mode working with custom hook
