+++ src/App.tsx
@@ -5,8 +5,8 @@ import { FormPanel } from "./components/FormPanel/FormPanel";
 import { CountriesList } from "./components/CountriesList/CountriesList";
 import { GlobalStyle } from "./GlobalStyles";
 import { CountriesDataContext } from "./Context/CountriesDataContext";
-import { ThemeChangeProvider, useThemeContext } from "./Context/ThemeContext";
 import { darkTheme } from "./darkTheme";
+import { useDarkThemeMode } from "./api/useDarkThemeMode";
 
 const StyledMain = styled.main`
 	display: flex;
@@ -15,20 +15,18 @@ const StyledMain = styled.main`
 `;
 
 export const App = () => {
-	const { isDarkTheme } = useThemeContext();
-	console.log(isDarkTheme);
+	const { isDarkTheme, toggleTheme } = useDarkThemeMode();
+
 	return (
-		<ThemeChangeProvider>
-			<CountriesDataContext>
-				<ThemeProvider theme={isDarkTheme ? darkTheme : lightTheme}>
-					<GlobalStyle />
-					<StyledMain>
-						<Header />
-						<FormPanel />
-						<CountriesList />
-					</StyledMain>
-				</ThemeProvider>
-			</CountriesDataContext>
-		</ThemeChangeProvider>
+		<CountriesDataContext>
+			<ThemeProvider theme={isDarkTheme ? darkTheme : lightTheme}>
+				<GlobalStyle />
+				<StyledMain>
+					<Header toggleTheme={toggleTheme} />
+					<FormPanel />
+					<CountriesList />
+				</StyledMain>
+			</ThemeProvider>
+		</CountriesDataContext>
 	);
 };
