+++ src/stories/Header.stories.tsx
@@ -1,8 +1,8 @@
-import { Header } from "../components/Header/Header";
+import { Header, HeaderProps } from "../components/Header/Header";
 
 export default {
 	title: "Header",
 	component: Header,
 };
 
-export const Default = () => <Header />;
+export const Default = (args: HeaderProps) => <Header {...args} />;
